refactor: :truck: move `check_data()` into root of folder
