+++ src/seedcase_sprout/core/__init__.py
@@ -10,6 +10,7 @@
 # from .delete_resource_properties import *
 
 from .as_readme_text import as_readme_text
+from .check_data import check_data
 from .create_resource_properties import create_resource_properties
 from .create_resource_structure import create_resource_structure
 from .examples import (
@@ -37,7 +38,6 @@ from .properties import (
 )
 from .read_properties import read_properties
 from .read_resource_batches import read_resource_batches
-from .sprout_checks.check_data import check_data
 from .sprout_checks.check_properties import (
     check_package_properties,
     check_properties,
