+++ src/seedcase_sprout/core/join_resource_batches.py
@@ -1,8 +1,8 @@
 import polars as pl
 
+from seedcase_sprout.core.check_data import check_data
 from seedcase_sprout.core.constants import BATCH_TIMESTAMP_COLUMN_NAME
 from seedcase_sprout.core.properties import ResourceProperties
-from seedcase_sprout.core.sprout_checks.check_data import check_data
 from seedcase_sprout.core.sprout_checks.check_properties import (
     check_resource_properties,
 )
