+++ src/seedcase_sprout/core/read_resource_batches.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import polars as pl
 
+from seedcase_sprout.core.check_data import check_data
 from seedcase_sprout.core.check_is_file import check_is_file
 from seedcase_sprout.core.constants import (
     BATCH_TIMESTAMP_COLUMN_NAME,
@@ -11,7 +12,6 @@ from seedcase_sprout.core.constants import (
     BATCH_TIMESTAMP_PATTERN,
 )
 from seedcase_sprout.core.properties import ResourceProperties
-from seedcase_sprout.core.sprout_checks.check_data import check_data
 from seedcase_sprout.core.sprout_checks.check_properties import (
     check_resource_properties,
 )
