+++ tests/core/sprout_checks/test_check_data.py
@@ -5,6 +5,10 @@ import polars as pl
 from polars.testing import assert_frame_equal
 from pytest import fixture, mark, raises
 
+from seedcase_sprout.core.check_data import (
+    _FRICTIONLESS_TO_ALLOWED_POLARS_TYPES,
+    check_data,
+)
 from seedcase_sprout.core.examples import (
     example_data,
     example_data_all_types,
@@ -15,10 +19,6 @@ from seedcase_sprout.core.properties import (
     FieldProperties,
     ResourceProperties,
     TableSchemaProperties,
-)
-from seedcase_sprout.core.sprout_checks.check_data import (
-    _FRICTIONLESS_TO_ALLOWED_POLARS_TYPES,
-    check_data,
 )
 from tests.core.assert_raises_errors import (
     assert_raises_check_errors,
